requirements.txt
-e .
pytest
aws_cdk.aws_lambda
aws_cdk.aws_apigateway